Adjust config file sources.
