+++ package.json
@@ -6,8 +6,8 @@
   "type": "module",
   "exports": {
     ".": "./dest/index.js",
-    "./eslint": "./estlint-config/index.cjs",
-    "./prettier": "./prettier-config/index.cjs",
+    "./eslint": "./.eslintrc.cjs",
+    "./prettier": "./.prettierrc.json",
     "./async-map": "./dest/async-map/index.js",
     "./aztec-address": "./dest/eth-address/index.js",
     "./crypto": "./dest/crypto/index.js",
