--- src/prettier-config/index.cjs
@@ -1,6 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  trailingComma: 'all',
-  printWidth: 120,
-  arrowParens: 'avoid',
-};
